Extract shared indexing utilities to src/indexing/utils.ts
Move globToRegex, matchesPatterns, hasLowSemanticValue to a shared
utils module. source-indexer.ts re-exports for backwards compat.
Direct consumers (bash-fs, test-path-filter) updated to import
from utils.ts.
