+++ src/mcp/tools/bash-fs.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { Bash } from 'just-bash';
-import { matchesPatterns } from '../../indexing/source-indexer.js';
+import { matchesPatterns } from '../../indexing/utils.js';
 import type { SourceConfig } from '../../types.js';
 import { generateIndexMd, generateSearchTipsMd } from './bash-virtual-files.js';
